+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheet Example/ExampleSwiftUIPaymentSheet.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2021 stripe-ios. All rights reserved.
 //
 
-@_spi(STP) import Stripe
+import Stripe
 import SwiftUI
 
 struct ExampleSwiftUIPaymentSheet: View {